Refactor code to use a types file
