+++ src/components/posts/EditableArticle.tsx
@@ -1,17 +1,17 @@
 import Article from "./Article";
 import { CardContent } from "@mui/material";
-import PostForm, { PostFormValues } from "./PostForm";
-import { ISimplePost } from "./Post";
+import PostForm from "./PostForm";
 import { AxiosResponse } from "axios";
+import { PostFormValues } from "../../types";
 
-interface IProps {
-    post: ISimplePost;
+interface Props {
+    post: PostFormValues;
     isEditing: boolean;
     onSubmit: (data: PostFormValues) => Promise<AxiosResponse>;
     onDiscard: () => void;
 }
 
-const EditableArticle = ({ post, isEditing, onSubmit, onDiscard }: IProps) => {
+const EditableArticle = ({ post, isEditing, onSubmit, onDiscard }: Props) => {
     return (
         <>
             {isEditing ? (
