+++ src/components/posts/PostList.tsx
@@ -1,15 +1,15 @@
-import { PostInterface } from "../../pages/PostPage.tsx";
+import { Post } from "../../types.ts";
 import PostPreview from "./PostPreview.tsx";
 import PostPreviewSkeleton from "./PostPreviewSkeleton.tsx";
 
-interface PostListInterface {
-    posts: PostInterface[];
+interface Props {
+    posts: Post[];
     isLoading: boolean;
 }
 
 const POSTS_LENGTH = 20;
 
-const PostList = ({ posts, isLoading }: PostListInterface) => {
+const PostList = ({ posts, isLoading }: Props) => {
     const renderPosts = isLoading
         ? Array.from({ length: POSTS_LENGTH }).map((_, idx) => (
               <PostPreviewSkeleton key={idx} />
