+++ src/constants.tsx
@@ -1,5 +1,7 @@
 export const ACCESS_TOKEN_STORAGE_ID = "LC-access";
 export const REFRESH_TOKEN_STORAGE_ID = "LC-refresh";
+export const NAV_WIDTH = 250;
+export const RANKINGS_LENGTH = 50;
 
 export const themeSettings = (mode: "light" | "dark") => {
     return {
@@ -17,3 +19,210 @@ export const themeSettings = (mode: "light" | "dark") => {
         },
     };
 };
+
+export const LANGUAGE_OPTIONS = [
+    ["aa", "Afar"],
+    ["ab", "Abkhazian"],
+    ["af", "Afrikaans"],
+    ["ak", "Akan"],
+    ["am", "Amharic"],
+    ["an", "Aragonese"],
+    ["ar", "Arabic"],
+    ["ar-dz", "Algerian Arabic"],
+    ["as", "Assamese"],
+    ["ast", "Asturian"],
+    ["av", "Avaric"],
+    ["az", "Azerbaijani"],
+    ["ba", "Bashkir"],
+    ["be", "Belarusian"],
+    ["bg", "Bulgarian"],
+    ["bm", "Bambara"],
+    ["bn", "Bengali"],
+    ["bo", "Tibetan"],
+    ["br", "Breton"],
+    ["bs", "Bosnian"],
+    ["ca", "Catalan"],
+    ["ce", "Chechen"],
+    ["co", "Corsican"],
+    ["cr", "Cree"],
+    ["cs", "Czech"],
+    ["cy", "Welsh"],
+    ["da", "Danish"],
+    ["de", "German"],
+    ["dsb", "Lower Sorbian"],
+    ["dz", "Dzongkha"],
+    ["ee", "Ewe"],
+    ["el", "Greek"],
+    ["en", "English"],
+    ["en-au", "Australian English"],
+    ["en-gb", "British English"],
+    ["eo", "Esperanto"],
+    ["es", "Spanish"],
+    ["es-ar", "Argentinian Spanish"],
+    ["es-co", "Colombian Spanish"],
+    ["es-mx", "Mexican Spanish"],
+    ["es-ni", "Nicaraguan Spanish"],
+    ["es-ve", "Venezuelan Spanish"],
+    ["et", "Estonian"],
+    ["eu", "Basque"],
+    ["fa", "Persian"],
+    ["ff", "Fulah"],
+    ["fi", "Finnish"],
+    ["fj", "Fijian"],
+    ["fo", "Faroese"],
+    ["fr", "French"],
+    ["fy", "Frisian"],
+    ["ga", "Irish"],
+    ["gd", "Scottish Gaelic"],
+    ["gl", "Galician"],
+    ["gn", "Guarani"],
+    ["grc", "Ancient Greek"],
+    ["gu", "Gujarati"],
+    ["gv", "Manx"],
+    ["ha", "Hausa"],
+    ["he", "Hebrew"],
+    ["hi", "Hindi"],
+    ["hr", "Croatian"],
+    ["hsb", "Upper Sorbian"],
+    ["ht", "Haitian"],
+    ["hu", "Hungarian"],
+    ["hy", "Armenian"],
+    ["hz", "Herero"],
+    ["ia", "Interlingua"],
+    ["id", "Indonesian"],
+    ["ig", "Igbo"],
+    ["ii", "Sichuan Yi"],
+    ["ik", "Inupiaq"],
+    ["io", "Ido"],
+    ["is", "Icelandic"],
+    ["it", "Italian"],
+    ["iu", "Inuktitut"],
+    ["ja", "Japanese"],
+    ["jv", "Javanese"],
+    ["ka", "Georgian"],
+    ["kab", "Kabyle"],
+    ["kg", "Kongo"],
+    ["ki", "Kikuyu"],
+    ["kj", "Kuanyama"],
+    ["kk", "Kazakh"],
+    ["kl", "Kalaallisut"],
+    ["km", "Khmer"],
+    ["kn", "Kannada"],
+    ["ko", "Korean"],
+    ["kr", "Kanuri"],
+    ["ks", "Kashmiri"],
+    ["ku", "Kurdish"],
+    ["kv", "Komi"],
+    ["ky", "Kyrgyz"],
+    ["la", "Latin"],
+    ["lb", "Luxembourgish"],
+    ["lg", "Ganda"],
+    ["li", "Limburgan"],
+    ["ln", "Lingala"],
+    ["lo", "Lao"],
+    ["lt", "Lithuanian"],
+    ["lu", "Luba-Katanga"],
+    ["lv", "Latvian"],
+    ["mg", "Malagasy"],
+    ["mh", "Marshallese"],
+    ["mi", "Maori"],
+    ["mk", "Macedonian"],
+    ["ml", "Malayalam"],
+    ["mn", "Mongolian"],
+    ["mr", "Marathi"],
+    ["ms", "Malay"],
+    ["mt", "Maltese"],
+    ["my", "Burmese"],
+    ["na", "Nauru"],
+    ["nb", "Norwegian Bokmål"],
+    ["nd", "North Ndebele"],
+    ["ne", "Nepali"],
+    ["ng", "Ndonga"],
+    ["nl", "Dutch"],
+    ["nn", "Norwegian Nynorsk"],
+    ["nr", "South Ndebele"],
+    ["nv", "Navajo"],
+    ["ny", "Chichewa"],
+    ["oc", "Occitan"],
+    ["oj", "Ojibwa"],
+    ["om", "Oromo"],
+    ["or", "Oriya"],
+    ["os", "Ossetic"],
+    ["pa", "Punjabi"],
+    ["phn", "Phoenician"],
+    ["pi", "Pali"],
+    ["pl", "Polish"],
+    ["ps", "Pashto"],
+    ["pt", "Portuguese"],
+    ["pt-br", "Brazilian Portuguese"],
+    ["qu", "Quechua"],
+    ["rm", "Romansh"],
+    ["rn", "Rundi"],
+    ["ro", "Romanian"],
+    ["ru", "Russian"],
+    ["rw", "Kinyarwanda"],
+    ["sa", "Sanskrit"],
+    ["sc", "Sardinian"],
+    ["sd", "Sindhi"],
+    ["se", "Northern Sami"],
+    ["sg", "Sango"],
+    ["si", "Sinhala"],
+    ["sk", "Slovak"],
+    ["sl", "Slovenian"],
+    ["sm", "Samoan"],
+    ["sn", "Shona"],
+    ["so", "Somali"],
+    ["sq", "Albanian"],
+    ["sr", "Serbian"],
+    ["sr-latn", "Serbian Latin"],
+    ["st", "Southern Sotho"],
+    ["su", "Sundanese"],
+    ["sv", "Swedish"],
+    ["sw", "Swahili"],
+    ["ta", "Tamil"],
+    ["te", "Telugu"],
+    ["tg", "Tajik"],
+    ["th", "Thai"],
+    ["ti", "Tigrinya"],
+    ["tk", "Turkmen"],
+    ["tl", "Tagalog"],
+    ["tn", "Tswana"],
+    ["to", "Tonga"],
+    ["tpi", "Tok Pisin"],
+    ["tr", "Turkish"],
+    ["ts", "Tsonga"],
+    ["tt", "Tatar"],
+    ["tw", "Twi"],
+    ["ty", "Tahitian"],
+    ["udm", "Udmurt"],
+    ["ug", "Uighur"],
+    ["uk", "Ukrainian"],
+    ["ur", "Urdu"],
+    ["uz", "Uzbek"],
+    ["ve", "Venda"],
+    ["vi", "Vietnamese"],
+    ["vo", "Volapük"],
+    ["wa", "Walloon"],
+    ["wo", "Wolof"],
+    ["xh", "Xhosa"],
+    ["yi", "Yiddish"],
+    ["yo", "Yoruba"],
+    ["za", "Zhuang"],
+    ["zh-hans", "Simplified Chinese"],
+    ["zh-hant", "Traditional Chinese"],
+    ["zu", "Zulu"],
+];
+export const LEVEL_OPTIONS = [
+    ["A1", "Beginner"],
+    ["A2", "Elementary"],
+    ["B1", "Intermediate"],
+    ["B2", "Upper-Intermediate"],
+    ["C1", "Advanced"],
+    ["C2", "Proficient"],
+];
+export const GENDER_OPTIONS = [
+    ["M", "Male"],
+    ["F", "Female"],
+    ["O", "Other"],
+    ["U", "Prefer not to say"],
+];
