+++ src/layouts/navbar/Navbar.tsx
@@ -12,25 +12,14 @@ import HomeIcon from "@mui/icons-material/Home";
 import EmojiEventsIcon from '@mui/icons-material/EmojiEvents';
 import ListItemText from "@mui/material/ListItemText";
 import Collapse from "@mui/material/Collapse";
+import { NAV_WIDTH } from "../../constants";
 
-export interface NavbarProps {
+interface Props {
     isNavOpen: boolean;
     onNavClose: () => void;
 }
 
-const NAV_WIDTH = 250;
-
-/**
- * Renders the navbar component
- *
- * Props:
- * - isNavOpen
- * - onNavClose
- *
- * {SimpleLayout} -> Navbar
- */
-
-const Navbar = ({ isNavOpen, onNavClose }: NavbarProps) => {
+const Navbar = ({ isNavOpen, onNavClose }: Props) => {
     const [open, setOpen] = useState(false);
 
     const handleClick = () => {
