+++ src/pages/RankingPage.tsx
@@ -3,7 +3,7 @@ import ListItem from "@mui/material/ListItem";
 import ListItemText from "@mui/material/ListItemText";
 import ListItemAvatar from "@mui/material/ListItemAvatar";
 import Avatar from "@mui/material/Avatar";
-import { Divider, SvgIconProps, Typography } from "@mui/material";
+import { Divider, Typography } from "@mui/material";
 import CheckCircleOutlineIcon from "@mui/icons-material/CheckCircleOutline";
 import ArticleIcon from "@mui/icons-material/Article";
 import AddIcon from "@mui/icons-material/Add";
@@ -12,20 +12,10 @@ import RankingService from "../service/ranking.service";
 import UserStat from "../components/rankings/UserStat";
 import React from "react";
 import RankingSkeleton from "../components/rankings/RankingSkeleton";
+import { Ranking, UserStat as IUserStat } from "../types";
+import { RANKINGS_LENGTH } from "../constants";
 
-interface IRanking {
-    username: string;
-    total: number;
-    posts: number;
-    corrections: number;
-    joined: string;
-}
-
-const userStats: Array<{
-    title: string;
-    icon: React.ReactElement<SvgIconProps>;
-    key: keyof IRanking;
-}> = [
+const userStats: Array<IUserStat> = [
     { title: "Total contributions", icon: <AddIcon />, key: "total" },
     { title: "Entries published", icon: <ArticleIcon />, key: "posts" },
     {
@@ -34,8 +24,6 @@ const userStats: Array<{
         key: "corrections",
     },
 ];
-
-const RANKINGS_LENGTH = 50;
 
 const RankingPage = () => {
     const { isLoading, isError, data } = useQuery({
@@ -47,7 +35,7 @@ const RankingPage = () => {
         ? Array.from({ length: RANKINGS_LENGTH }).map((_, idx) => (
               <RankingSkeleton key={idx} />
           ))
-        : data.map((ranking: IRanking) => (
+        : data.map((ranking: Ranking) => (
               <React.Fragment key={ranking.username}>
                   <ListItem key={ranking.username}>
                       <ListItemAvatar>
