+++ src/service/auth.service.tsx
@@ -1,21 +1,12 @@
 import api from "./api";
+import { UserRegisterFormValues } from "../types";
 
-export interface IRegisterUser {
-    username: string;
-    password: string;
-    password2: string;
-    email: string;
-    native_language: string;
-    studying_language: string;
-    studying_level: string;
-    gender: string;
-}
 
 const login = async (username: string, password: string) => {
     return await api.post(`/token/`, { username, password });
 };
 
-const register = async (data: IRegisterUser) => {
+const register = async (data: UserRegisterFormValues) => {
     return await api.post(`users/~create`, data);
 };
